chore(tests): move helpers to `__helpers__` directory
