+++ tests/__helpers__/regexEscape.js
